src/services/tweet_capture/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.18.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
